imagePicker/requirements.txt
Django==1.9.4
djangorestframework
django-solo
markdown
django-filter
qiniu
numpy
psycopg2
git+https://github.com/InstaSketch/algolib.git@develop
